[TASK] Enable PHP 7.4 support for composer-based installations
Currently, TYPO3 Core (8.7 only) restricts PHP
versions to PHP 7.0-7.3.

In order to enable testing and projects running
PHP 7.4 the restriction needs to be lifted.

As we expect PHP 7.4 to be the last major PHP version in
the PHP v7 series, we use ^7.0 now.

Please note that testing framework dependencies
still need some modifications in order to allow
PHP 7.4, however having this restriction in place
fatals in various places currently, so it should
be removed.

Used composer command:
  composer req "php:^7.0"

Resolves: #89906
Releases: 8.7
Change-Id: Iff89166c56f82c9c28114686492e24a7049bc88e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62591
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>