[!!!][TASK] Return moved PID when overlaying versioned records
In the past, the API methods:
* PageRepository->versionOL()
* BackendUtility::workspaceOL()

returned the PID of the moved record in a workspace in the field
_ORIG_PID and the live record value in "pid".

The methods "fixVersioningPid()" were used to return the
moved PID (of the versioned records) and set the "_ORIG_PID"
to the value of the live record.

This inconsistency can be overcome by simply using workspaceOL
returning the newly moved location as "pid" and keep the "_ORIG_pid"
of the live record for later purposes.

This way, fixVersioningPid is not needed anymore, if workspaceOL/versionOL
is applied. TYPO3 Core is not using the fixVersioningPid methods
anymore now and they are marked as deprecated.

As this is a breaking change ("behavior of overlay and value of _ORIG_pid")
an RST explains the new behavior.

Resolves: #92598
Releases: master
Change-Id: I618d282d490b386e1a75caae3ab657f1605e3cb3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66169
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>