[TASK] Use multi record selection in recycler module
The multi record selection component is now also used
in the recycler module. This increases consistency, as
custom checkbox logic could be removed.

Also the recyclers' JavaScript implementation of the
selection actions is simplified. To interact with the
multi record selection component, some events are used.
This also means, the functionality stays the same, while
most of it is now handled by the Multi record selection
component.

Resolves: #94926
Releases: master
Change-Id: I57d91d070bfa75c01d4cb87dfd0c5576d111dcfa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70678
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>