[TASK] Align scheduler add/edit form of a task with FormEngine markup
With #100042 the markup of the form fields in the add/edit form of a
task has been updated. With this patch the markup was updated again
to fully match the FormEngine's form field markup.
With this we create the prerequisite that rollouts of global style
adjustments for form fields also work on this form.

Mid-term, this module should use FormEngine for rendering anyway,
so this is kind of another pre-patch.

Resolves: #100601
Related: #100042
Related: #100726
Releases: main
Change-Id: I4ec2f7a1607bfb5005610cf2ab3cd5c2a2aac851
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78643
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>