[BUGFIX] Prevent recycler failure when allowDelete=0 is set
Setting the user TSconfig option `mod.recycler.allowDelete=0`
lead to the recycler's backend module showing a javascript
error that an undefined HTML DOM element (mass delete button)
cannot be removed.

The check whether the mass delete should be created in the
DOM is performed in the fluid section `multiRecordSelectionActions`
of `RecyclerModule.html`. Because of this, there is no
need to try to remove this in the DOM dynamically.

The whole conditional block is now removed, because the other
functionality it provided was to set a `display: block`
attribute on the mass delete buttons, which is also already part of
the regular DOM.

Two other checks on the existance of the DOM mass delete element
are now changed so that `TYPO3.settings.Recycler.deleteDisable` is
evaluated instead.

Resolves: #105883
Releases: main, 13.4
Change-Id: Ia32bfedd0e26c5e9a294a4c923de69316bc48efe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88173
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>